--- data/totalhp_ru.json
@@ -1,47 +1,0 @@
-{
-    "hp_panel": {
-        "x": 0,
-        "y": 30,
-        "width": 186,
-        "height": 24,
-        "hcentered": true,
-        "vcentered": false,
-        "background": "totalhp_bg.dds",
-        "font": "default_medium.font",
-        "visible": true
-    },
-    "maincaliber_panel": {
-        "x": 150,
-        "y": 30,
-        "width": 160,
-        "height": 15,
-        "hcentered": true,
-        "vcentered": false,
-        "background": "totalhp_bg.dds",
-        "font": "default_smaller.font",
-        "text": "Основной калибр: ",
-        "color": "#EEEEEE",
-        "visible": true
-    },
-    "avgdamage_panel": {
-        "x": -150,
-        "y": 30,
-        "width": 160,
-        "height": 15,
-        "hcentered": true,
-        "vcentered": false,
-        "background": "totalhp_bg.dds",
-        "font": "default_smaller.font",
-        "text": "Средний урон: ",
-        "color": "#EEEEEE",
-        "visible": true
-    },
-    "colors": [
-            {"value": 0.67, "color": "#FE0E00"},
-            {"value": 0.95, "color": "#FFFFFF"},
-            {"value": 1.05, "color": "#FFFFFF"},
-            {"value": 1.5, "color": "#60FF00"}
-    ],
-    "team_damage": "<font color='#00EAFF'>{{attacker-name}} ({{attacker-vehicle}})</font> повредил союзника {{victim-name}} ({{victim-vehicle}}) на <font color='#FFDD33'>{{damage}}</font>!",
-    "show_team_damage": true
-}
